+++ test/color_picker_action_buttons_test.dart
@@ -82,6 +82,7 @@ void main() {
       dialogCancelButtonType: ColorPickerActionButtonType.outlined,
       dialogOkButtonLabel: 'OKAY',
       dialogOkButtonType: ColorPickerActionButtonType.elevated,
+      useRootNavigator: false,
     );
 
     test(
@@ -89,7 +90,7 @@ void main() {
         'EXPECT exact print string value.', () {
       expect(m4.toString(), equalsIgnoringHashCodes(
           // ignore: lines_longer_than_80_chars
-          'ColorPickerActionButtons#00000(okButton: true, closeButton: true, okIcon: IconData(U+0E085), closeIcon: IconData(U+0E139), closeIsLast: false, okTooltip: "GO", closeTooltip: "OFF", closeTooltipIsClose: false, toolIconsThemeData: IconThemeData#00000(color: Color(0xff000000), opacity: 1.0, size: 30.0), visualDensity: VisualDensity#00000(h: -1.0, v: -1.0)(horizontal: -1.0, vertical: -1.0), alignment: Alignment.topLeft, splashRadius: 30.0, constraints: BoxConstraints(46.0<=w<=Infinity, 46.0<=h<=Infinity), dialogActionButtons: false, dialogActionIcons: true, dialogCancelButtonLabel: "DONE", dialogCancelButtonType: outlined, dialogOkButtonLabel: "OKAY", dialogOkButtonType: elevated)'));
+          'ColorPickerActionButtons#00000(okButton: true, closeButton: true, okIcon: IconData(U+0E085), closeIcon: IconData(U+0E139), closeIsLast: false, okTooltip: "GO", closeTooltip: "OFF", closeTooltipIsClose: false, toolIconsThemeData: IconThemeData#00000(color: Color(0xff000000), opacity: 1.0, size: 30.0), visualDensity: VisualDensity#00000(h: -1.0, v: -1.0)(horizontal: -1.0, vertical: -1.0), alignment: Alignment.topLeft, splashRadius: 30.0, constraints: BoxConstraints(46.0<=w<=Infinity, 46.0<=h<=Infinity), dialogActionButtons: false, dialogActionIcons: true, dialogCancelButtonLabel: "DONE", dialogCancelButtonType: outlined, dialogOkButtonLabel: "OKAY", dialogOkButtonType: elevated, useRootNavigator: false)'));
     });
     test(
         'CPAB2.2: Test toStringShort implemented via debugFillProperties '
@@ -136,6 +137,7 @@ void main() {
           dialogCancelButtonType: ColorPickerActionButtonType.outlined,
           dialogOkButtonLabel: 'OKAY',
           dialogOkButtonType: ColorPickerActionButtonType.elevated,
+          useRootNavigator: false,
         ),
         equals(m4),
       );
@@ -171,6 +173,7 @@ void main() {
             dialogCancelButtonType: null,
             dialogOkButtonLabel: null,
             dialogOkButtonType: null,
+            useRootNavigator: null,
           ),
           equals(m2));
     });
@@ -200,6 +203,7 @@ void main() {
           dialogCancelButtonType: ColorPickerActionButtonType.outlined,
           dialogOkButtonLabel: 'OKAY',
           dialogOkButtonType: ColorPickerActionButtonType.elevated,
+          useRootNavigator: false,
         ),
         equals(m4),
       );